Normalizing the State shape
