+++ src/reducers/index.js
@@ -3,7 +3,7 @@ import {combineReducers} from 'redux';
 import todos, * as fromTodos from './todos'
 
 const todoApp = combineReducers({
-    todos,
+    todos
 });
 
 export default todoApp;
