+++ src/reducers/todos.js
@@ -1,51 +1,49 @@
-const todo = (state, action) => {
+import {combineReducers} from 'redux';
+
+import todo from './todo';
+
+const byId = (state = {}, action) => {
     switch (action.type) {
         case 'ADD_TODO':
-            return {
-                id: action.id,
-                text: action.text,
-                completed: false
-            };
         case 'TOGGLE_TODO':
-            if (state.id !== action.id) {
-                return state;
-            }
             return {
                 ...state,
-                completed: !state.completed
+                [action.id]: todo(state[action.id], action),
             };
-
         default:
             return state;
     }
 };
 
-const todos = (state = [], action) => {
+const allIds = (state = [], action) => {
     switch (action.type) {
         case 'ADD_TODO':
-            return [
-                ...state,
-                todo(undefined, action)
-            ];
-        case 'TOGGLE_TODO':
-            return state.map(x => todo(x, action));
+            return [...state, action.id];
         default:
             return state;
     }
 };
+
+const todos = combineReducers({
+    byId,
+    allIds
+});
 
 export default todos;
 
+const getAllTodos = (state) => state.allIds.map(id => state.byId[id]);
+
 export const getVisibleTodos = (state, filter) => {
+    const allTodos = getAllTodos(state);
     switch (filter) {
         case 'all':
-            return state;
+            return allTodos;
         case 'completed':
-            return state.filter(
+            return allTodos.filter(
                 t => t.completed
             );
         case 'active':
-            return state.filter(
+            return allTodos.filter(
                 t => !t.completed
             );
         default:
